Big chase game
This is a game in which four players compete with each other for food which appers randomly on the screen. every time a player eats food it's strength increase and can eliminate a player of lower strength by touching it. last player who survives is declared winner. It's a simple game made in java.